# job-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme Emploi",
        "description": "Documentation Swagger des endpoints Admin pour la gestion des utilisateurs, offres et candidatures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Lister les offres favorites du candidat connecté",
                "operationId": "de458f646b8ab1d9eb829bbd6a218808",
                "responses": {
                    "200": {
                        "description": "Liste paginée des offres favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Favorite"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/{id}/favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Ajouter ou retirer une offre d'emploi des favoris",
                "operationId": "cb48051500480d6878946624694bc472",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut du favori modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job added to favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre d'emploi non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/{id}/favorite/check": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Vérifier si une offre est dans les favoris de l'utilisateur",
                "operationId": "728d64e17285ff750505c020242b4b7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 8
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultat de la vérification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_favorite": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Favorite": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "job_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "job": {
                        "$ref": "#/components/schemas/Job"
                    }
                },
                "type": "object"
            },
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Laravel Senior"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du poste..."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris, France"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 55000
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "full_time",
                            "part_time",
                            "contract",
                            "internship"
                        ],
                        "example": "full_time"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "technology",
                            "healthcare",
                            "education",
                            "finance",
                            "other"
                        ],
                        "example": "technology"
                    },
                    "application_deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}